✨ Feature/33 archive user (#45)
✨ Archive User
✨ Fail login & forgot/reset password of inactive user
🐛 Fix User Update if no team ids are passed in payload

+++ api/src/services/users/users.ts
@@ -63,9 +63,11 @@ export const updateUser: MutationResolvers['updateUser'] = async ({
       update: {},
     })
   }
-  await db.membership.deleteMany({
-    where: { userId: user.id, NOT: { teamId: { in: teamIds } } },
-  })
+  if (teamIds) {
+    await db.membership.deleteMany({
+      where: { userId: user.id, NOT: { teamId: { in: teamIds } } },
+    })
+  }
   return user
 }
